deduplicate tar junk in index.py (move to utils)
